--- src/features/onboarding/components/buttons/WelcomePageButton/index.ts
@@ -1,1 +1,0 @@
-export { WelcomePageButton } from './WelcomePageButton'
